feature/commande-view : migration com_ttc, calcult de com_ht et com_ttc à l'ajout d'un nouveau produit dans le panier

+++ models/commande.js
@@ -76,7 +76,7 @@ module.exports = (sequelize, DataTypes) => {
       com_tva: DataTypes.DECIMAL,
       com_port: DataTypes.FLOAT,
       com_frais: DataTypes.FLOAT,
-      com_ttc: DataTypes.DECIMAL,
+      com_ttc: DataTypes.FLOAT,
       com_infos: DataTypes.STRING,
       com_comment: DataTypes.TEXT,
       com_adr_liv: DataTypes.INTEGER,
